src/main/java/org/craftedsw/harddependencies/TripService.java
package org.craftedsw.harddependencies;

import java.util.ArrayList;
import java.util.List;

import org.craftedsw.harddependencies.exception.UserNotLoggedInException;
import org.craftedsw.harddependencies.trip.Trip;
import org.craftedsw.harddependencies.trip.TripDAO;
import org.craftedsw.harddependencies.user.User;
import org.craftedsw.harddependencies.user.UserSession;

public class TripService {

	public List<Trip> getTripsBy(User friend, User loggedInUser) throws UserNotLoggedInException {
		validate(loggedInUser);
		return friend.isFriendsWith(loggedInUser) ? tripsBy(friend) : noTrips();
	}

	private ArrayList<Trip> noTrips() {
		return new ArrayList<Trip>();
	}

	private User validate(User loggedInUser) throws UserNotLoggedInException {
		if (loggedInUser == null) {
			throw new UserNotLoggedInException();
		}
		return loggedInUser;
	}

	protected List<Trip> tripsBy(User user) {
		return TripDAO.findTripsByUser(user);
	}
}
